# optimize rx operator