--- a/operators/rx.go
+++ b/operators/rx.go
@@ -28,21 +28,19 @@ func newRX(options rules.OperatorOptions) (rules.Operator, error) {
 }
 
 func (o *rx) Evaluate(tx rules.TransactionState, value string) bool {
-	match := o.re.FindStringSubmatch(value)
-	if len(match) == 0 {
-		return false
-	}
 
 	if tx.Capturing() {
+		match := o.re.FindStringSubmatch(value)
+		if len(match) == 0 {
+			return false
+		}
 		for i, c := range match {
-			if i == 9 {
-				return true
-			}
 			tx.CaptureField(i, c)
 		}
+		return true
+	} else {
+		return o.re.MatchString(value)
 	}
-
-	return true
 }
 
 func init() {