There is no need to find all submatches if there is no capture.

```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/operators
BenchmarkRxSubstringVsMatch/Find_all_RX-10         	1000000000	         0.0000116 ns/op	       0 B/op	       0 allocs/op
BenchmarkRxSubstringVsMatch/Find_only_first-10     	1000000000	         0.0000119 ns/op	       0 B/op	       0 allocs/op
BenchmarkRxSubstringVsMatch/Find_only_N-10         	1000000000	         0.0000145 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/corazawaf/coraza/v3/operators	0.666s

```